# GH-96421: Introduce cleanup contexts for inlined frames