This is intended as an alternative to #96319. I personally find that this is a bit simpler, more scalable, and easier to reason about.

Rather than pushing "shim" frames, give `_PyInterpreterFrame` a `cleanup` member that tells the interpreter how to handle its two exit paths (the old `exit_unwind` label on error, and the new `cleanup` label on success).

This proof-of-concept sketches out how inlining generators and class constructions would work under this new scheme by implementing their cleanup code in `exit_unwind` and `cleanup`. However, the specializations themselves are still unimplemented.

<!-- gh-issue-number: gh-96421 -->
* Issue: gh-96421
<!-- /gh-issue-number -->