@markshannon 2022-10-28
I don't see how this is simpler.

Once you strip out the change note and documentation, the PRs are about the same size.
Half of code added in https://github.com/python/cpython/pull/96319 is for creating these sort of trampoline functions. Even if we use the approach in this PR, it just isn't as general as the inserting shim functions, so we will need shim functions anyway.
In terms of changes to the interpreter this PR is much more complex.

How the cleanup contexts compose? How do you do two things in a cleanup context?
For example, since having to pop a chunk is when popping a frame is so rare, it might make sense to insert a shim frame to handle that (it probably wouldn't, it's just an example). 
How would cleanup contexts handle that use case and another another special case?