@brandtbucher 2022-10-31
> I don't see how this is simpler.

What I find simpler is that all of the new complexity is limited to a very specific part of `_PyEval_EvalFrameDefault`, rather than than *everywhere frames are used*. It also maintains a simple, intuitive rule that has served us well thus far: one frame per Python call in the source program. It also avoids the overhead of pushing and initializing additional frames in lots of places, which seems like a real cost (reflected in the benchmark results), not to mention the burden of maintaining hand-written bytecode, location tables, etc.

> Once you strip out the change note and documentation, the PRs are about the same size. Half of code added in #96319 is for creating these sort of trampoline functions.

Well, that's a bit unfair. The majority of this PR is the PoC implementations of generator and class inlining, which I thought you might like to see. I've removed those parts since the point's been made, and the PR is downright *tiny* now. ;)

Also, this PR doesn't need as much documentation, since it doesn't break out-of-process debuggers and profilers like the other PR (which I see as a strong point in favor of this approach).

> Even if we use the approach in this PR, it just isn't as general as the inserting shim functions, so we will need shim functions anyway.

For what? I've demonstrated that inlining generators and class construction is quite straightforward with this approach. To me, it seems general enough for other places where we might want to insert arbitrary cleanup code between frames.

I could be missing something, though. What else do we "need" shim functions for?

> In terms of changes to the interpreter this PR is much more complex.

If by "interpreter" you mean "the eval loop", then *maybe*. But if by "interpreter" you mean "the entire runtime", then I disagree (strongly).

> How the cleanup contexts compose? How do you do two things in a cleanup context? For example, since having to pop a chunk is when popping a frame is so rare, it might make sense to insert a shim frame to handle that (it probably wouldn't, it's just an example). How would cleanup contexts handle that use case and another another special case?

If we really needed something like this (it sounds like we might not, though), I imagine it could be quite straightforward to do using flags for composable handlers rather than enumerated values, as I've done here.

Coming out of the 3.11 prereleases, I feel quite strongly that the last things we need in 3.12 are gratuitous third-party breakage and additional (delicate) rules about how frames should be handled. Why not start simple by leaving shim frames out entirely, then add all of that only when we have a justified need? I'm not 100% opposed to shim frames, I just think that an approach like this one is a better way of solving the same set of problems right now.